Big changes to the way background images are generated from wire (or wire-like) data.
Instead of storing in 2D vectors and later converting to numpy, the data
is stored in a 1D C-style format (though still as vector) and a numpy 2D
array is initialized in C++ to from the data buffer holding the data.

Therefore, the processing has been trimmed to:
-Read data
-Subtract Pedestal
-Display data
Without any copying or conversion of items (and especially none in python)

This should give a sizeable performance boost to microboone.